+++ src/main/java/repository/ThemeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import model.Theme;
 
 
-public interface ThemeRepository extends CrudRepository<Theme, Long>{
+public interface ThemeRepository extends CrudRepository<Theme, Integer>{
